requirements.txt
flask
requests
google-cloud-secret-manager
google-cloud-documentai
#google-api-python-client
